feat: PR#16レビュー対応④ — BottomNav を左サイドバーに変更
- BottomNav（ボトムナビ）を廃止し Sidebar（左サイドバー）を新規作成
- サイドバーに「Strength Log」ロゴ + ナビリンク4件（アイコン付き）を配置
- Layout を flex-row 化し、サイドバー右側にヘッダー＋メインコンテンツを配置
- BottomNav ディレクトリ削除、BottomNavProps → SidebarProps リネーム
- テスト 283/283 PASS
