--- src/components/BottomNav/BottomNav.module.css
@@ -1,49 +1,0 @@
-.nav {
-  display: flex;
-  position: fixed;
-  bottom: 0;
-  left: 50%;
-  transform: translateX(-50%);
-  width: 100%;
-  max-width: 480px;
-  background: linear-gradient(180deg, #0c1e3d 0%, #1565c0 100%);
-  border-top: 1px solid rgba(100, 181, 246, 0.25);
-  z-index: 100;
-}
-
-.tab {
-  flex: 1;
-  text-decoration: none;
-  display: flex;
-  justify-content: center;
-  padding: 0.5rem 0;
-}
-
-.tabInner {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 2px;
-}
-
-.icon {
-  font-size: 1.25rem;
-}
-
-.label {
-  font-size: 0.65rem;
-  color: rgba(255, 255, 255, 0.45);
-}
-
-.tab[data-active='true'] .label {
-  color: #90caf9;
-  font-weight: 600;
-}
-
-.tabInner[data-active='true'] {
-  color: #90caf9;
-}
-
-.tabInner[data-active='false'] {
-  color: rgba(255, 255, 255, 0.45);
-}
